validation: use zero objects as fallback
When we evaluate rules, we need to check if the JSONPath we get
in the `Path` rules exists.
However, VM objects have a lot of optional paths and subpaths,
and because no JSONPath library can (apparently?) distinguish
between missing key and illegal path, a not-fully-initialized
legal path may be misdetected as illegal path.

Furthermore, if a legal path leads to a value which is optional,
and not specified we should evaluate the rule with the zero value.

To solve both issues, we introduce a new helper package which
gives us fully initialized zero valued VM objects.

This way we can make the validation in two steps:
1. we look for the path in the object. If it's there, we're done
   and we can go ahead
2. if it is not there, we try again in the complete, zero-valued
   object. If it's there, we are done. Otherwise, the path is
   considered illegal.

Signed-off-by: Francesco Romani <fromani@redhat.com>