flags: propagate the flag values to klog
klog is a glog fork introduced (apparently? didn't dug deeply)
into k8s.io/client-go >= 10.0.0

We need to keep in sync all the flags values we use between
glog, klog (both using goflag) and the main module (using pflag).

This patch uses a mixed approach. We add the flag set to the pflag
flag set like the doc suggests for glog integration, and we borrow the
explicit loop cdi/apiserver is using to deal with this case.

Signed-off-by: Francesco Romani <fromani@redhat.com>