travis: fetch third party data earlier
To test the upgrade path, we need the kubevirt release info
from github.
In the last few weeks, it happened few times that the fetch
operation failed late in the build (we used to do the fetch just in
time), wasting a lot of time.

We move the fetch of the github-dependent info as early as possible
in the build, and we cache locally. This way, if the fetch fails,
we know early and we waste as little as possible.

Having the versions info cached allows us also to simplify
some upgrade steps.

Signed-off-by: Francesco Romani <fromani@redhat.com>